x5

+++ app/src/main/java/appmoviles/com/appsmoviles20191/db/DBHandler.java
@@ -86,7 +86,7 @@ public class DBHandler extends SQLiteOpenHelper {
 
     public void deleteAmigosofUser(String uid) {
         SQLiteDatabase db = this.getWritableDatabase();
-        db.execSQL("DELETE FROM "+TABLE_AMIGOS + " WHERE "+AMIGOS_USERID+"=''");
+        db.execSQL("DELETE FROM "+TABLE_AMIGOS + " WHERE "+AMIGOS_USERID+"='"+uid+"'");
         db.close();
     }
 }
